backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlparse
xarray
netCDF4
pandas
sentence-transformers
chromadb
openai
python-dateutil
python-multipart
aiofiles
plotly
